Big Idea: Payload Structs.
For "simplifying" the PlaceItems part. Would not require seperate CursorStates for what items the cursor is holding because the struct carries big time (handles placing items in there). this means more structs, but less code in CursorInfo.cs which might not be something bad.
sleep about it. decide on it another day.